backend/requirements.txt
fastapi==0.118.0
uvicorn[standard]==0.37.0
sqlmodel==0.0.24
pydantic-settings==2.11.0
python-dotenv==1.1.1
pytest==8.4.2
httpx==0.28.1
